feat!: Add lobby state to example and fix player ordering
This changes the API of the new methods so they're async. I needed to do
this because there issues with connecting a new WsStream inside the
frankenstein bevy polling mechanism. So this lets it connect first,
outside the message loop.

Since we now have states, and bevy starts immediately, I was able to
finally remove the super ugly polling hack at the start, though, and
with it a lot of annoying dependencies.