Fix incorrect usage of .fuse() within select! macro
.fuse() construct a FusedFuture from a regular future, the goal being to
prevent the original future from being polled after completion, but when
it is called from within the select! macro, we just create a new
FusedFuture each time, which meant that we might very well poll it after
completion as new wrappers were created all the time.

I'm not really sure what the consequences of this would be though, but
better to be on the safe side.