ci: Run github actions ci workflow on pull_request events
We want to run github actions on pull requests, not just on pushes,
because the pull_request event actually performs the merge to the target
branch.

This means if a branch started on an old main, we will still catch
potential incompatible changes that don't show up as merge conflicts.

Keeping the push events as well for now, as public github actions is
free for public repos. If this changes, we should probably remove the
push events except on main and tags.