Replace Warp with Axum for the signalling server (#82)
It's no secret Warp is losing the server framework battle to Axum, so this will
be the first step towards modernizing the architecture.

This replaces Warp with Axum for the signalling server. Another opinion I made
was to use tracing/tracing_subscriber over pretty_env_logger for logging, for
similar reasons.

This is a total drop-in replacement. All metrics being logged and all route
rules are preserved, with no impact to demos or the user.

Co-authored-by: Johan Klokkhammer Helsing <johanhelsing@gmail.com>